rename sampler.run to sampler.run_sample
